WebCL-conformance/conformance/bindingTesting/queues/cl_commandQueue_enqueueWaitForEvents.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueWaitForEvents conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    debug("WebCLCommandQueue :: enqueueWaitForEvents() test case.");
    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var webCLEvent = wtu.createEvent();
    var srcBuffer = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, BUFFER_SIZE);
    var dstBuffer = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, BUFFER_SIZE);

    wtu.enqueueCopyBuffer(webCLCommandQueue, srcBuffer, dstBuffer, 0, 0, BUFFER_SIZE, [], webCLEvent);
    shouldBeUndefined("webCLCommandQueue.enqueueWaitForEvents([webCLEvent]);");

    var webCLUserEvent = wtu.createUserEvent(webCLContext);
    //shouldBeUndefined("webCLCommandQueue.enqueueWaitForEvents([webCLUserEvent]);");

    // Testing for Exceptions.
    shouldThrowExceptionName("webCLCommandQueue.enqueueWaitForEvents([]);", "INVALID_VALUE");

    var releasedEvent = wtu.createUserEvent(webCLContext);
    wtu.release(releasedEvent);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWaitForEvents([releasedEvent]);", "INVALID_EVENT_WAIT_LIST");

    // Testing for exception "INVALID_CONTEXT"
    var webCLContext1 = wtu.createContext();
    var webCLCommandQueue1 = wtu.createCommandQueue(webCLContext1);
    var webCLUserEvent1 = wtu.createUserEvent(webCLContext1);
    var srcBuffer1 = wtu.createBuffer(webCLContext1, webcl.MEM_READ_ONLY, BUFFER_SIZE);
    var dstBuffer1 = wtu.createBuffer(webCLContext1, webcl.MEM_READ_ONLY, BUFFER_SIZE);

    var webCLContext2 = wtu.createContext();
    var webCLCommandQueue2 = wtu.createCommandQueue(webCLContext2);
    var webCLUserEvent2 = wtu.createUserEvent(webCLContext2);
    var srcBuffer2 = wtu.createBuffer(webCLContext2, webcl.MEM_READ_ONLY, BUFFER_SIZE);
    var dstBuffer2 = wtu.createBuffer(webCLContext2, webcl.MEM_READ_ONLY, BUFFER_SIZE);

    shouldThrowExceptionName("webCLCommandQueue1.enqueueWaitForEvents([webCLUserEvent2]);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue2.enqueueWaitForEvents([webCLUserEvent1]);", "INVALID_CONTEXT");
    wtu.setStatus(webCLUserEvent1, webcl.COMPLETE);
    wtu.setStatus(webCLUserEvent2, webcl.COMPLETE);

    var webCLEvent1 = wtu.createEvent();
    wtu.enqueueCopyBuffer(webCLCommandQueue1, srcBuffer1, dstBuffer1, 0, 0, BUFFER_SIZE, [], webCLEvent1);
    shouldThrowExceptionName("webCLCommandQueue2.enqueueWaitForEvents([webCLEvent1]);", "INVALID_CONTEXT");
    var webCLEvent2 = wtu.createEvent();
    wtu.enqueueCopyBuffer(webCLCommandQueue2, srcBuffer2, dstBuffer2, 0, 0, BUFFER_SIZE, [], webCLEvent2);
    shouldThrowExceptionName("webCLCommandQueue1.enqueueWaitForEvents([webCLEvent2]);", "INVALID_CONTEXT");

} catch (e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
